+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Features:
   - Update Tests to use Admin API to create preconditions -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/100)
+  - Dashboard: My open Issues -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/74)
+  - Cannot open external issue from Modal and View of Issue -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/96)
+  - Test run Bulk Delete -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/91)
 
 Bugfixes:
   - Test run View Page performance is bad -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/99)